requirements.txt
streamlit
opencv-python
numpy
matplotlib
Pillow
streamlit-image-comparison
